+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### [Current changes]
 - **[Add]** Add capability for handling alert as desired.
+- **[Add]** Add action read scanned pdf
 
 ### [14.11.2][March 30, 2022]
 - **[Fix]** Pyautogui import fix for ui-less machine